+++ tests/resources/tf/bug67_missingDefinitionOnAttribute.tf
@@ -1,0 +1,5 @@
+resource "aws_elb" "aws_elb_1" {
+    listener = {
+        instance_port = "1"
+    }
+}